requirements.txt
attrs==17.2.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
backports.weakref==1.0.post1
bash==0.6
bleach==1.5.0
bokeh==0.12.5
certifi==2016.2.28
click==6.7
configparser==3.5.0
contextlib2==0.5.5
cycler==0.10.0
decorator==4.1.2
docopt==0.6.2
eli5==0.8
entrypoints==0.2.3
enum34==1.1.6
Flask==0.12
Flask-Cors==3.0.3
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gevent==1.2.2
git-review==1.25.0
gitdb==0.6.4
GitPython==2.0.8
graphviz==0.8
greenlet==0.4.12
h5py==2.7.1
html5lib==0.9999999
humanize==0.5.1
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.1
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
Keras==2.0.8
keras-vis==0.4.1
kitchen==1.2.4
lime==0.1.1.24
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mock==2.0.0
more-itertools==2.2
mpmath==0.19
nbconvert==5.3.1
nbformat==4.4.0
neptune-cli==2.0.0
networkx==2.0
notebook==5.1.0
numpy==1.13.1
oauthlib==2.0.4
olefile==0.44
pandas==0.20.3
pandocfilters==1.4.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.4.0
psutil==4.3.0
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.2
pykwalify==1.5.2
pyparsing==2.2.0
python-dateutil==2.4.2
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.11
pyzmq==16.0.2
quiver-engine==0.1.4.1.4
raven==6.1.0
requests==2.11.1
requests-oauthlib==0.8.0
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
smmap==0.9.0
subprocess32==3.2.7
sympy==1.1.1
tabulate==0.7.7
tensorflow==1.3.0
tensorflow-gpu==1.1.0
tensorflow-tensorboard==0.1.6
terminado==0.6
terminaltables==2.1.0
testpath==0.3.1
tornado==4.5.2
tqdm==4.11.2
traitlets==4.3.2
typing==3.6.2
urllib3==1.16
voluptuous==0.9.3
wcwidth==0.1.7
websocket-client==0.35.0
Werkzeug==0.12.2
widgetsnbextension==3.0.3
